+++ NEWS.md
@@ -2,6 +2,7 @@
 ## unreleased
 - Use dpkg options --force-confdef and --force-confold when installing
   packages using the apt provider
+- Add OpenBSD package provider
 
 ## 0.5.0 (2015-09-21)
 - Add attribute hash support to DSL and resources